Add iSCSI to nova-compute AppArmor profile
When nova-compute is deployed with AppArmor in enforce mode, it
fails to attach iSCSI volumes because it is being blocked.

This patch fixes that by updating the nova-compute AppArmor profile
to include the commands, paths and capabilities used to attach
iSCSI volumes.

Change-Id: I46b8a8453cdc045f10850958652af920c8a19660
Closes-bug: #1815519
(cherry picked from commit ad8b38884e63ec31852031980dca94977a42b315)